added db connection and clean up
