+++ pom.xml
@@ -97,6 +97,19 @@
       <artifactId>springdoc-openapi-ui</artifactId>
       <version>1.6.4</version>
     </dependency>
+
+    <!-- email service -->
+    <dependency>
+      <groupId>org.springframework</groupId>
+      <artifactId>spring-context-support</artifactId>
+      <version>5.3.25</version>
+    </dependency>
+    <dependency>
+      <groupId>com.sun.mail</groupId>
+      <artifactId>jakarta.mail</artifactId>
+      <version>1.6.7</version>
+    </dependency>
+
   </dependencies>
 
   <build>
