+++ src/main/java/com/team3/central/services/EmailService.java
@@ -1,5 +1,30 @@
 package com.team3.central.services;
 
-public interface EmailService {
-    void sendSimpleMessage(String to, String subject, String text);
+import org.springframework.beans.factory.annotation.Autowired;
+
+import org.springframework.mail.MailException;
+import org.springframework.mail.SimpleMailMessage;
+import org.springframework.mail.javamail.JavaMailSender;
+import org.springframework.stereotype.Service;
+
+@Service
+public class EmailService {
+
+    private JavaMailSender emailSender;
+
+    @Autowired
+    public EmailService(JavaMailSender emailSender){
+        this.emailSender=emailSender;
+    }
+    public void sendSimpleMessage(String to, String subject, String text) {
+        try {
+            SimpleMailMessage message = new SimpleMailMessage();
+            message.setTo(to);
+            message.setSubject(subject);
+            message.setText(text);
+            emailSender.send(message);
+        } catch (MailException exception) {
+            exception.printStackTrace();
+        }
+    }
 }
