--- src/main/java/com/team3/central/services/EmailServiceImpl.java
@@ -1,30 +1,0 @@
-package com.team3.central.services;
-
-import org.springframework.beans.factory.annotation.Autowired;
-
-import org.springframework.mail.MailException;
-import org.springframework.mail.SimpleMailMessage;
-import org.springframework.mail.javamail.JavaMailSender;
-import org.springframework.stereotype.Service;
-
-@Service("EmailService")
-public class EmailServiceImpl implements EmailService {
-
-    private JavaMailSender emailSender;
-
-    @Autowired
-    public EmailServiceImpl(JavaMailSender emailSender){
-        this.emailSender=emailSender;
-    }
-    public void sendSimpleMessage(String to, String subject, String text) {
-        try {
-            SimpleMailMessage message = new SimpleMailMessage();
-            message.setTo(to);
-            message.setSubject(subject);
-            message.setText(text);
-            emailSender.send(message);
-        } catch (MailException exception) {
-            exception.printStackTrace();
-        }
-    }
-}
