+++ src/main/resources/application.yml
@@ -18,3 +18,15 @@ spring:
         dialect: org.hibernate.dialect.PostgresPlusDialect
     hibernate:
       ddl-auto: update
+# Email service
+  mail:
+    host: ${MAIL_HOST}
+    port: 587
+    username: ${MAIL_USERNAME}
+    password: ${MAIL_PASSWORD}
+    properties:
+      mail:
+        smtp:
+          auth: true
+          starttls:
+            enable: true
