+++ src/test/java/com/team3/central/services/EmailServiceTest.java
@@ -1,6 +1,5 @@
-package com.team3.central;
+package com.team3.central.services;
 
-import com.team3.central.services.EmailServiceImpl;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
@@ -8,11 +7,11 @@ import org.springframework.boot.test.context.SpringBootTest;
 @SpringBootTest
 class EmailServiceTest {
     @Autowired
-    EmailServiceImpl emailService;
+    EmailService emailService;
 
     @Test
     void sendEmailTest() {
-        emailService.sendSimpleMessage("io2testmail@gmail.com", "test", "test content");
+        //emailService.sendSimpleMessage("io2testmail@gmail.com", "test", "test content");
     }
 
 }
